+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `isSelected` prop to `ControlBarButton` component
 - Add `UpAndDownCaret` icon component
 - Add channel list component for chat
+- Add `children` prop for ControlBarItem to allow for custom PopOver content
 
 ### Changed